Program requirements
Hello everyone, I’m have a hard time attempting to decrypt a message
here on my program. So far I am able to export an encrypted file, but
can’t seem to figure out how to decrypt or decode a txt file on here.
Any help would be great.

Morse Code uses a series of short and long pulses called dots and
dashes, respectively, to encode letters and digits. For example, the
letter A is “dot-dash,” B is “dash-dot-dot- dot.” Find a table of Morse
Code on the Internet to create a dictionary for mapping characters and
digits to Morse Code. Use the dictionary to create a program that can
translate between Morse Code and characters and digits.

Complete Chapter 9, exercise 21, with the following modifications: the
program is to calculate error detection values for a message and
encrypt both the message and error detection codes into Morse-code. The
message will be sent to another student via e-mail as a .txt file
attachment. The received e-mail file (.txt file attachment) will be
unencrypted and checked using the error detection code. Thus, your
program must be able to encrypt and decrypt Morse-code, while
performing the error detection.

Results to be submitted included both plaintext and Morse-code version
for both sent mail and received mail. Message content is to include
your first and last name; and the two well-known Morse-code messages.
The first is the message Samuel Morse used to the demonstration the
telegraph to Congress in 1844, "What hath God wrought", which is based
on the Bible scripture reference in Numbers 23:23. The second is the
internationally recognized distress call “SOS”, which is a commonly
understood acronym based on the phase “Save Our Souls”.  Example:

Steve Powelson
What hath God wrought
SOS

The error detection algorithm will be provided by the instructor.